handle SIP being enabled

+++ lib/tasks/ntfs.sh
@@ -3,10 +3,14 @@
 # NTFS-3G functions
 
 ntfs_prepare () {
-  if [ -f /sbin/mount_ntfs ]; then
-    sudo mv /sbin/mount_ntfs /sbin/mount_ntfs.original
+  if csrutil status | grep "System Integrity Protection status: disabled"; then
+    echo "(SIP Enabled)" >&3
+
+    if [ -f /sbin/mount_ntfs ]; then
+      sudo mv /sbin/mount_ntfs /sbin/mount_ntfs.original
+    fi
+    sudo ln -sf /usr/local/sbin/mount_ntfs /sbin/mount_ntfs
   fi
-  sudo ln -sf /usr/local/sbin/mount_ntfs /sbin/mount_ntfs
 
   my_sleep "${@}"
 }
